+++ tests/tests_integration/test_api/test_geospatial.py
@@ -19,6 +19,7 @@ from cognite.client.data_classes.geospatial import (
     PropertyAndSearchSpec,
 )
 from cognite.client.exceptions import CogniteAPIError
+from tests.utils import set_request_limit
 
 FIXED_SRID = 121111 + random.randint(0, 1_000)
 
@@ -449,6 +450,16 @@ class TestGeospatialAPI:
         feature_list = FeatureList(list(features))
         assert len(feature_list) == len(many_features)
 
+    def test_list(self, cognite_client, test_feature_type, test_features):
+        with set_request_limit(cognite_client.geospatial, 2):
+            res = cognite_client.geospatial.list_features(
+                feature_type_external_id=test_feature_type.external_id, properties={"externalId": {}}, limit=4
+            )
+
+        assert len(res) == 4
+        df = res.to_pandas()
+        assert list(df) == ["externalId"]
+
     def test_to_pandas(self, test_feature_type, test_features):
         df = test_features.to_pandas(camel_case=True)
         assert list(df) == [
@@ -559,3 +570,22 @@ class TestGeospatialAPI:
             group_by=["externalId"],
         )
         assert len(res) == 4
+
+    def test_aggregate_with_order_by(self, cognite_client, test_feature_type, test_features):
+        res = cognite_client.geospatial.aggregate_features(
+            feature_type_external_id=test_feature_type.external_id,
+            property="temperature",
+            aggregates=["count"],
+            group_by=["externalId"],
+            order_by=[OrderSpec("externalId", "DESC")],
+        )
+        external_ids = [item.external_id for item in res]
+        external_ids.reverse()
+        res_asc = cognite_client.geospatial.aggregate_features(
+            feature_type_external_id=test_feature_type.external_id,
+            property="temperature",
+            aggregates=["count"],
+            group_by=["externalId"],
+            order_by=[OrderSpec("externalId", "ASC")],
+        )
+        assert external_ids == [item.external_id for item in res_asc]